refactor(navigation): extract student routes
